// lib/form-field-logic/types.ts
export type OrgShortName = "kapiti" | "wellington";

// FORM OPTIONS TYPES
// Generated by https://quicktype.io

export interface FormDetail {
	kapiti: Forms;
	wellington: Forms;
}

export interface Forms {
	clientDetails: Details;
	jobDetails: Details;
}

export interface Details {
	restfulName: FormRestfulName;
	structure: Structure[];
	fields: FieldMap; // string is Field.fieldName value
}

export type FieldMap = { [Key in FieldNameKey]: Field };

export type FieldNameKey = `${"cd" | "jd"}_${FieldName}`;

export interface Structure {
	sectionTitle: string;
	fields: FieldName[][];
}

export type FieldName = string;

export type FormRestfulName = "clients" | "jobs";

export interface Field {
	label: string;
	fieldName: FieldName;
	fieldType: FieldType;
	values?: (number | string)[];
	readOnly?: boolean;
	isRequired?: boolean;
	submitExclude?: boolean;
	additionalClasses?: AdditionalClass[] | string;
	validationDetail?: string;
}

export type FieldType =
	| "input"
	| "hidden"
	| "select"
	| "datalist"
	| "Text"
	| "textarea"
	| "tel"
	| "date"
	| "number";

export type AdditionalClass = "account-name" | "d-none" | "staff-hours";

// NEW FORMS DATA STRUCTURE

export type FormName = "client" | "job" | "staff";

export interface NewFormsObject {
	client: NewFormDetail;
	job: NewFormDetail;
	staff: StaffFormDetail;
}

export type ClientFormFieldName = string; // to be replaced by union type with all values included

export type JobFormFieldName = string; // to be replaced by union type with all values included

export interface NewFormDetail {
	restfulName: FormRestfulName;
	sections: Section[];
}

export interface Section {
	sectionTitle: string;
	rows: NewField[][];
}

export interface NewField {
	label: string;
	name: string;
	type: FieldType;
	values?: (number | string)[];
	properties: FieldProperties;
	classes?: FormFieldClass;
}

export interface FieldProperties {
	readOnly?: boolean;
	isRequired?: boolean;
	excludeOnSubmit?: boolean; // called 'submitExclude' in Field object
}

export type FormFieldClass = string[] | string;

export interface StaffFormDetail {}

export type StaffName = string;